Java language level migrations (Refactoring, non-functional changes)
git-svn-id: https://svn.l3s.uni-hannover.de/learnweb/trunk/LearnwebOER/Learnweb@5882 422a2c9f-66af-4668-9993-60dc3b0871b8

+++ src/de/l3s/learnweb/forum/ForumTopicsBean.java
@@ -46,7 +46,7 @@ public class ForumTopicsBean extends ApplicationBean implements Serializable
             {
                 Integer id = getParameterInt("group_id");
                 if(id != null)
-                    groupId = id.intValue();
+                    groupId = id;
             }
 
             group = getLearnweb().getGroupManager().getGroupById(groupId);
